// .test-build/const/types.d.ts
export declare enum SyncMode {
    /** 同步执行 */
    Sync = 0,
    /** 异步执行 */
    Async = 1
}